Change SentimentAnalysis feature to extend Feature
Since the original implementation was built before the `Feature` class
existed, the function version had to be ported over to the new format.

This is functionally identical to the previous implementation, swapping
the original function for the `SentimentAnalysis#score` method.

Tests have also been updated to use the new class, and moved into the
`tests` directory with the rest of them.
