+++ features/sentiment_analysis.py
@@ -1,12 +1,15 @@
 from nltk.sentiment.vader import SentimentIntensityAnalyzer
+from model.feature import Feature
 
 
 intensityAnalyzer = None
 
 
-def sentiment_analysis(text):
-    global intensityAnalyzer
-    if intensityAnalyzer is None:
-        intensityAnalyzer = SentimentIntensityAnalyzer()
-    scores = intensityAnalyzer.polarity_scores(text)
-    return scores['compound']
+class SentimentAnalysis(Feature):
+
+    def score(self, text):
+        global intensityAnalyzer
+        if intensityAnalyzer is None:
+            intensityAnalyzer = SentimentIntensityAnalyzer()
+        scores = intensityAnalyzer.polarity_scores(text)
+        return scores['compound']
